Move generic test utils for evc to testing module
